Move event group validator to TimeScout.Application
