+++ TimeScout.Application/Validator/EventGroup/EventGroupCreateRequestValidator.cs
@@ -1,8 +1,7 @@
-using System;
 using FluentValidation;
-using TimeScout.API.DTOs.EventGroup;
+using TimeScout.Application.DTOs.EventGroup;
 
-namespace TimeScout.API.Validator.Event;
+namespace TimeScout.Application.Validator.EventGroup;
 
 public class EventGroupCreateRequestValidator : AbstractValidator<EventGroupCreateRequestDto>
 {
